requirements.txt
python-dotenv
black
jupyterlab
dspy
openai
ujson